parseInt
How parseInt works, this is a method to convert string to integer number with a base octal, decimal o binary, and it's the reason we used 10, after the input. To indicate to convert to base 10 = decimal.
This broken up, since only use the valid data, and with * this was broken.